# internal/gateway/server/server.go
package server

import (
	"context"
	"net/http"
	
	"github.com/gin-gonic/gin"
	"github.com/prometheus/client_golang/prometheus/promhttp"
	"github.com/company/mindgateway/internal/gateway/handlers"
	"github.com/company/mindgateway/internal/shared/config"
	"github.com/sirupsen/logrus"
)

type Server struct {
	router *gin.Engine
	config *config.Config
	logger *logrus.Logger
	
	// Modular components
	authClient     AuthClient
	registryClient RegistryClient
	routingEngine  RoutingEngine
	queueManager   QueueManager
}

type Option func(*Server)

func New(opts ...Option) (*Server, error) {
	s := &Server{
		router: gin.New(),
	}
	
	for _, opt := range opts {
		opt(s)
	}
	
	s.setupRoutes()
	s.setupMiddleware()
	
	return s, nil
}

func (s *Server) setupRoutes() {
	// Health checks
	s.router.GET("/health", handlers.Health())
	s.router.GET("/ready", handlers.Ready())
	
	// Metrics
	s.router.GET("/metrics", gin.WrapH(promhttp.Handler()))
	
	// API routes
	v1 := s.router.Group("/v1")
	{
		// OpenAI compatible endpoints
		v1.POST("/chat/completions", s.handleChatCompletion)
		v1.POST("/completions", s.handleCompletion)
		v1.POST("/embeddings", s.handleEmbeddings)
	}
	
	// Admin routes
	admin := s.router.Group("/admin")
	admin.Use(s.requireAdmin())
	{
		admin.GET("/workers", s.listWorkers)
		admin.GET("/queue", s.queueStatus)
	}
}

func (s *Server) Start() error {
	return s.router.Run(s.config.Server.Address)
}

func (s *Server) Shutdown(ctx context.Context) error {
	// Graceful shutdown logic
	return nil
}

// Option functions
func WithConfig(cfg *config.Config) Option {
	return func(s *Server) {
		s.config = cfg
	}
}

func WithLogger(logger *logrus.Logger) Option {
	return func(s *Server) {
		s.logger = logger
	}
}

// Handler methods
func (s *Server) handleChatCompletion(c *gin.Context) {
	// TODO: Implement
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func (s *Server) handleCompletion(c *gin.Context) {
	// TODO: Implement
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func (s *Server) handleEmbeddings(c *gin.Context) {
	// TODO: Implement
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func (s *Server) listWorkers(c *gin.Context) {
	// TODO: Implement
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func (s *Server) queueStatus(c *gin.Context) {
	// TODO: Implement
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func (s *Server) requireAdmin() gin.HandlerFunc {
	return func(c *gin.Context) {
		// TODO: Implement admin check
		c.Next()
	}
}

func (s *Server) setupMiddleware() {
	// Use common middleware
	s.router.Use(gin.Recovery())
	s.router.Use(gin.Logger())
	// Add custom middleware here
}

// Interface definitions for modular components
type AuthClient interface {
	ValidateToken(ctx context.Context, token string) (bool, error)
	GetUserRoles(ctx context.Context, userID string) ([]string, error)
}

type RegistryClient interface {
	GetActiveWorkers(ctx context.Context) ([]Worker, error)
}

type RoutingEngine interface {
	RouteRequest(ctx context.Context, req interface{}) (Worker, error)
}

type QueueManager interface {
	Enqueue(ctx context.Context, req interface{}, priority int) (string, error)
	Dequeue(ctx context.Context) (interface{}, error)
}

type Worker struct {
	ID       string
	Name     string
	Endpoint string
	Models   []string
	Load     float64
	Status   string
}